Move functionality to specify service endpoint location to the bus
